+++ moves.py
@@ -40,13 +40,34 @@ def possible_positions_hr(piece,row,col):
 
 # POSSIBLE HORIZONTAL LEFT POSITIONS TO MOVE A PIECE TO:
 
-def possible_positions_hl(rook,row,col):
+def possible_positions_hl(piece,row,col):
     possibles=[]
     for next_col in range(col-1,-1,-1):
-        other_piece = rook.__board__.get_piece(row,next_col)
+        other_piece = piece.__board__.get_piece(row,next_col)
         if other_piece is not None:
-            if other_piece.__color__ != rook.__color__:
+            if other_piece.__color__ != piece.__color__:
                 possibles.append((row,next_col))  
             break  
         else:possibles.append((row,next_col))
     return possibles
+
+# POSSIBLE DIAGONAL ASCENDANT TO THE RIGHT POSITIONS TO A PIECE TO:
+
+def possible_positions_dar(piece,row,col):
+    pass
+
+# POSSIBLE DIAGONAL ASCENDANT TO THE LEFT POSITIONS TO A PIECE TO:
+
+def possible_positions_dal(piece,row,col):
+    pass
+
+# POSSIBLE DIAGONAL DESCENDANT TO THE RIGHT POSITIONS TO A PIECE TO:
+
+def possible_positions_ddr(piece,row,col):
+    pass
+
+# POSSIBLE DIAGONAL DESCENDANT TO THE LEFT POSITIONS TO A PIECE TO:
+
+def possible_positions_ddl(piece,row,col):
+    pass
+