+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.8] - 2024-10-01
+
+### Added
+
+- Tests for the rook creation 
+- Tests for bishop moves, str, valid_positions and vertical and horizontal wrong moves
+
 ## [0.1.7] - 2024-09-27
 
 ### Added